fix: download models on start up instead of request based.
